fix: incompatible ints

+++ graph/reader.c
@@ -100,7 +100,7 @@
 /* New (larger) length of a Point array, as function of the old; used when
    reallocating due to exhaustion of storage. */
 #define NEW_POINT_ARRAY_LENGTH(old_len) \
-((old_len)*sizeof(Point) < 10000000 ? 2 * (old_len) : (old_len) + 10000000/sizeof(Point))
+((old_len)*sizeof(Point) < 10000000 ? 2 * (old_len) : (old_len) + 10000000/(int)sizeof(Point))
 
 struct ReaderStruct		/* point reader datatype */
 {
